+++ src/main/vo/RepositoryRateVO.java
@@ -3,13 +3,13 @@ package main.vo;
 import java.util.Map;
 
 public class RepositoryRateVO {
-	private Map<String,Integer> rates;
+	private Map<String,Double> rates;
 
-	public Map<String, Integer> getRates() {
+	public Map<String, Double> getRates() {
 		return rates;
 	}
 
-	public void setRates(Map<String, Integer> rates) {
+	public void setRates(Map<String, Double> rates) {
 		this.rates = rates;
 	}
